+++ tests/testthat/test-zeroinfl.R
@@ -1,8 +1,6 @@
 if (require("testthat") &&
   require("insight") &&
   require("pscl")) {
-  context("insight, model_info")
-
   data("bioChemists")
 
   m1 <- zeroinfl(art ~ fem + mar + kid5 + ment | kid5 + phd, data = bioChemists)
@@ -110,5 +108,34 @@ if (require("testthat") &&
 
   test_that("find_statistic", {
     expect_identical(find_statistic(m1), "z-statistic")
+  })
+
+  test_that("get_statistic", {
+    expect_equal(
+      get_statistic(m1)$Statistic,
+      c(8.26297, -3.90986, 2.07134, -3.43156, 10.05389, -2.143, 0.21384, -1.84259),
+      tolerance = 1e-3
+    )
+    expect_equal(
+      get_statistic(m1)$Component,
+      c("conditional", "conditional", "conditional", "conditional",
+        "conditional", "zero_inflated", "zero_inflated", "zero_inflated"),
+      tolerance = 1e-3
+    )
+  })
+
+  m2 <- zeroinfl(formula = art ~ . | 1, data = bioChemists, dist = "negbin")
+  test_that("get_statistic", {
+    expect_equal(
+      get_statistic(m2)$Statistic,
+      c(1.8486, -2.97799, 1.83288, -3.32486, 0.42372, 8.38053, -0.21241),
+      tolerance = 1e-3
+    )
+    expect_equal(
+      get_statistic(m2)$Component,
+      c("conditional", "conditional", "conditional", "conditional",
+        "conditional", "conditional", "zero_inflated"),
+      tolerance = 1e-3
+    )
   })
 }